Store rocket's secret key in PEM format
