+++ src/login/mod.rs
@@ -21,8 +21,8 @@ pub(crate) use auto_login::*;
 use serde::Serialize;
 use time::{Duration, OffsetDateTime};
 mod code;
-mod keys;
-pub(crate) use keys::*;
+mod secret_key;
+pub(crate) use secret_key::*;
 mod redirect;
 pub(crate) use redirect::*;
 mod sudo;
