+++ src/main.rs
@@ -63,8 +63,8 @@ fn rocket() -> _ {
 }
 
 fn figment() -> Figment {
-    let keys = login::GameNightKeys::read_or_generate().unwrap();
-    Config::figment().merge(("secret_key", &keys.rocket_secret_key))
+    let secret_key = login::RocketSecretKey::read_or_generate().unwrap();
+    Config::figment().merge((rocket::Config::SECRET_KEY, &secret_key.0))
 }
 
 #[cfg(debug_assertions)]
